Move controllers into specific directories
Scoped all the routes under a corresponding module that allowed me to
move the controllers into specific directories for better organization
