--- app/controllers/api/v1/customer_favorite_merchant_controller.rb
@@ -1,10 +1,0 @@
-module Api
-  module V1
-    class CustomerFavoriteMerchantController < ApiController
-      def show
-        merchant_id = Customer.favorite_merchant(params[:id])
-        respond_with Merchant.find(merchant_id)
-      end
-    end
-  end
-end
