--- app/controllers/api/v1/customer_invoices_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class CustomerInvoicesController < ApiController
-      def index
-        respond_with Customer.find(params[:id]).invoices
-      end
-    end
-  end
-end
