--- app/controllers/api/v1/customer_transactions_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class CustomerTransactionsController < ApiController
-      def index
-        respond_with Invoice.transactions(params[:id])
-      end
-    end
-  end
-end
