+++ app/controllers/api/v1/customers/customer_invoices_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Customers
+      class CustomerInvoicesController < ApiController
+        def index
+          respond_with Customer.find(params[:id]).invoices
+        end
+      end
+    end
+  end
+end
