+++ app/controllers/api/v1/customers/customer_transactions_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Customers
+      class CustomerTransactionsController < ApiController
+        def index
+          respond_with Invoice.transactions(params[:id])
+        end
+      end
+    end
+  end
+end
