--- app/controllers/api/v1/invoice_customer_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class InvoiceCustomerController < ApiController
-      def show
-        respond_with Invoice.find(params[:id]).customer
-      end
-    end
-  end
-end
