--- app/controllers/api/v1/invoice_invoice_items_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class InvoiceInvoiceItemsController < ApiController
-      def index
-        respond_with Invoice.find(params[:id]).invoice_items
-      end
-    end
-  end
-end
