--- app/controllers/api/v1/invoice_item_invoice_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class InvoiceItemInvoiceController < ApiController
-      def show
-        respond_with InvoiceItem.find(params[:id]).invoice
-      end
-    end
-  end
-end
