--- app/controllers/api/v1/invoice_item_item_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class InvoiceItemItemController < ApiController
-      def show
-        respond_with InvoiceItem.find(params[:id]).item
-      end
-    end
-  end
-end
