+++ app/controllers/api/v1/invoice_items/invoice_item_item_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module InvoiceItems
+      class InvoiceItemItemController < ApiController
+        def show
+          respond_with InvoiceItem.find(params[:id]).item
+        end
+      end
+    end
+  end
+end
