+++ app/controllers/api/v1/invoice_items/invoice_items_controller.rb
@@ -1,0 +1,15 @@
+module Api
+  module V1
+    module InvoiceItems
+      class InvoiceItemsController < ApiController
+        def index
+          respond_with InvoiceItem.all
+        end
+
+        def show
+          respond_with InvoiceItem.find(params[:id])
+        end
+      end
+    end
+  end
+end
