--- app/controllers/api/v1/invoice_items_controller.rb
@@ -1,13 +1,0 @@
-module Api
-  module V1
-    class InvoiceItemsController < ApiController
-      def index
-        respond_with InvoiceItem.all
-      end
-
-      def show
-        respond_with InvoiceItem.find(params[:id])
-      end
-    end
-  end
-end
