--- app/controllers/api/v1/invoice_merchant_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class InvoiceMerchantController < ApiController
-      def show
-        respond_with Invoice.find(params[:id]).merchant
-      end
-    end
-  end
-end
