--- app/controllers/api/v1/invoice_transactions_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class InvoiceTransactionsController < ApiController
-      def index
-        respond_with Invoice.find(params[:id]).transactions
-      end
-    end
-  end
-end
