+++ app/controllers/api/v1/invoices/invoice_customer_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Invoices
+      class InvoiceCustomerController < ApiController
+        def show
+          respond_with Invoice.find(params[:id]).customer
+        end
+      end
+    end
+  end
+end
