+++ app/controllers/api/v1/invoices/invoice_invoice_items_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Invoices
+      class InvoiceInvoiceItemsController < ApiController
+        def index
+          respond_with Invoice.find(params[:id]).invoice_items
+        end
+      end
+    end
+  end
+end
