+++ app/controllers/api/v1/invoices/invoice_merchant_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Invoices
+      class InvoiceMerchantController < ApiController
+        def show
+          respond_with Invoice.find(params[:id]).merchant
+        end
+      end
+    end
+  end
+end
