+++ app/controllers/api/v1/invoices/invoice_transactions_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Invoices
+      class InvoiceTransactionsController < ApiController
+        def index
+          respond_with Invoice.find(params[:id]).transactions
+        end
+      end
+    end
+  end
+end
