--- app/controllers/api/v1/invoices_controller.rb
@@ -1,35 +1,0 @@
-module Api
-  module V1
-    class InvoicesController < ApiController
-      def index
-        respond_with Invoice.all
-      end
-
-      def show
-        respond_with Invoice.find(params[:id])
-      end
-
-      def find
-        respond_with Invoice.find_by(invoice_params)
-      end
-
-      def find_all
-        respond_with Invoice.where(invoice_params)
-      end
-
-      def random
-        respond_with Invoice.random
-      end
-
-      def items
-        respond_with Invoice.find(params[:id]).items
-      end
-
-      private
-
-      def invoice_params
-        params.permit(:id, :customer_id, :merchant_id, :status, :created_at, :updated_at)
-      end
-    end
-  end
-end
