--- app/controllers/api/v1/item_invoice_items_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class ItemInvoiceItemsController < ApiController
-      def index
-        respond_with Item.find(params[:id]).invoice_items
-      end
-    end
-  end
-end
