--- app/controllers/api/v1/item_merchant_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class ItemMerchantController < ApiController
-      def show
-        respond_with Item.find(params[:id]).merchant
-      end
-    end
-  end
-end
