+++ app/controllers/api/v1/items/item_invoice_items_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Items
+      class ItemInvoiceItemsController < ApiController
+        def index
+          respond_with Item.find(params[:id]).invoice_items
+        end
+      end
+    end
+  end
+end
