+++ app/controllers/api/v1/items/item_merchant_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Items
+      class ItemMerchantController < ApiController
+        def show
+          respond_with Item.find(params[:id]).merchant
+        end
+      end
+    end
+  end
+end
