+++ app/controllers/api/v1/items/items_controller.rb
@@ -1,0 +1,15 @@
+module Api
+  module V1
+    module Items
+      class ItemsController < ApiController
+        def index
+          respond_with Item.all
+        end
+
+        def show
+          respond_with Item.find(params[:id])
+        end
+      end
+    end
+  end
+end
