--- app/controllers/api/v1/merchant_invoices_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class MerchantInvoicesController < ApiController
-      def index
-        respond_with Merchant.find(params[:id]).invoices
-      end
-    end
-  end
-end
