--- app/controllers/api/v1/merchant_items_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class MerchantItemsController < ApiController
-      def index
-        respond_with Merchant.find(params[:id]).items
-      end
-    end
-  end
-end
