+++ app/controllers/api/v1/merchants/merchant_favorite_customer_controller.rb
@@ -1,0 +1,12 @@
+module Api
+  module V1
+    module Merchants
+      class MerchantFavoriteCustomerController < ApiController
+        def show
+          customer_id = Invoice.merchant_favorite_customer(params[:id])
+          respond_with Customer.find(customer_id)
+        end
+      end
+    end
+  end
+end
