+++ app/controllers/api/v1/merchants/merchant_invoices_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Merchants
+      class MerchantInvoicesController < ApiController
+        def index
+          respond_with Merchant.find(params[:id]).invoices
+        end
+      end
+    end
+  end
+end
