+++ app/controllers/api/v1/merchants/merchant_items_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Merchants
+      class MerchantItemsController < ApiController
+        def index
+          respond_with Merchant.find(params[:id]).items
+        end
+      end
+    end
+  end
+end
