--- app/controllers/api/v1/transaction_invoice_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  module V1
-    class TransactionInvoiceController < ApiController
-      def show
-        respond_with Transaction.find(params[:id]).invoice
-      end
-    end
-  end
-end
