+++ app/controllers/api/v1/transactions/transaction_invoice_controller.rb
@@ -1,0 +1,11 @@
+module Api
+  module V1
+    module Transactions
+      class TransactionInvoiceController < ApiController
+        def show
+          respond_with Transaction.find(params[:id]).invoice
+        end
+      end
+    end
+  end
+end
