+++ app/controllers/api/v1/transactions/transactions_finder_controller.rb
@@ -1,0 +1,25 @@
+module Api
+  module V1
+    module Transactions
+      class TransactionsFinderController < ApiController
+        def index
+          respond_with Transaction.where(transaction_params)
+        end
+
+        def show
+          respond_with Transaction.find_by(transaction_params)
+        end
+
+        def random
+          respond_with Transaction.random
+        end
+
+        private
+
+        def transaction_params
+          params.permit(:id, :invoice_id, :credit_card_number, :result, :created_at, :updated_at)
+        end
+      end
+    end
+  end
+end
