+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
 require "simplecov"
-SimpleCov.start
+SimpleCov.start "rails"
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
